fix: bug into auth routes
