+++ src/presentation/containers/post/Posts.container.tsx
@@ -9,6 +9,7 @@ import { ToastContainer } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css';
 import useNotification from '../../../hooks/common/useNotification'
 import { useAuth } from '../../../hooks/auth/useAuth'
+import useProfile from '../../../hooks/profile/useProfile'
 
 const initialValues:IPostFormValues = {
   title: '',
@@ -27,6 +28,7 @@ type Post = {
 const PostContainer: React.FC<Post> = ({ closeModal }: Post) => {
   // Hooks
   const { user } = useAuth()
+  const { profile } = useProfile()
   const { isLoading: posting, newPost } = usePost()
   const { isLoading: uploading, uploadFiles } = useUploadFile()
   const {successMessage, errorMessage} = useNotification()
@@ -110,7 +112,8 @@ const PostContainer: React.FC<Post> = ({ closeModal }: Post) => {
           id: user.uid,
           email: user.email || '',
           name: user.displayName || '',
-          avatar: user.photoUrl || ''
+          avatar: user.photoUrl || '',
+          profileUrl: profile.profile_url
         }
         
         const post = Object.assign(formData, {user: postUser, urls: urls})
